+++ admin/src/App.tsx
@@ -15,7 +15,6 @@ import { FinanceSales } from './pages/admin/FinanceSales';
 import { FinanceSettlement } from './pages/admin/FinanceSettlement';
 import { FinanceRefund } from './pages/admin/FinanceRefund';
 import { ManualMatching } from './pages/admin/ManualMatching';
-import { MatchingAlgorithm } from './pages/admin/MatchingAlgorithm';
 import { RecommendMatching } from './pages/admin/RecommendMatching';
 import { ReservationStatus } from './pages/admin/ReservationStatus';
 import { StatMatching } from './pages/admin/StatMatching';
@@ -61,7 +60,6 @@ function App() {
               <Route path="support/customer" element={<CustomerSupport />} />
               <Route path="support/manager" element={<ManagerSupport />} />
               <Route path="matching/manual" element={<ManualMatching />} />
-              <Route path="matching/algorithm" element={<MatchingAlgorithm />} />
               <Route path="matching/recommend" element={<RecommendMatching />} />
               <Route path="matching/reservation" element={<ReservationStatus />} />
               <Route path="statistics" element={<Navigate to="/admin/statistics/matching" replace />} />
