+++ admin/src/pages/admin/index.ts
@@ -5,13 +5,16 @@ export { Dashboard } from './Dashboard';
 export { UsersCustomer } from './UsersCustomer';
 export { UsersManager } from './UsersManager';
 export { UsersWaiting } from './UsersWaiting';
-export { default as UsersBlacklist } from './UsersBlacklist';
 export { CustomerSupport } from './CustomerSupport';
 export { ManagerSupport } from './ManagerSupport';
 export { ManualMatching } from './ManualMatching';
-export { MatchingAlgorithm } from './MatchingAlgorithm';
 export { RecommendMatching } from './RecommendMatching';
 export { ReservationStatus } from './ReservationStatus';
 export { FinanceSales } from './FinanceSales';
 export { FinanceSettlement } from './FinanceSettlement';
-export { FinanceRefund } from './FinanceRefund'; 
+export { FinanceRefund } from './FinanceRefund';
+export { StatMatching } from './StatMatching';
+export { StatReservation } from './StatReservation';
+export { StatSatisfaction } from './StatSatisfaction';
+export { StatRefund } from './StatRefund';
+export { default as UsersBlacklist } from './UsersBlacklist';
